--- lib/local_authority_data_importer.rb
@@ -1,81 +1,0 @@
-require 'csv'
-require 'redis'
-require 'redis-lock'
-require_relative 'redis_config'
-
-class LocalAuthorityDataImporter
-  include RedisConfig
-
-  def self.update_all
-    redis.lock("publisher:#{Rails.env}:local_authority_data_importer_lock", :life => 2.hours) do
-      begin
-        LocalServiceImporter.update
-        LocalInteractionImporter.update
-        LocalContactImporter.update
-        nagios_check(true, "Import succeeded")
-      rescue Exception => e
-        nagios_check(false, e.to_s)
-
-        # re-raise the exception so that it ends up in the logs
-        # and we allow through SyntaxError, Interrupt and SignalException
-        raise e
-      end
-    end
-  rescue Redis::Lock::LockNotAcquired => e
-    Rails.logger.debug("Failed to get lock for local directgov importing (#{e.message}). Another process probably got there first.")
-
-    # Flag nagios that this servers instance succeeded to stop lingering failures
-    nagios_check(true, "Importer not run on this instance")
-  end
-
-  def self.redis
-    Redis.new(REDIS_CONFIG)
-  end
-
-  def self.nagios_check(success, message)
-    code = success ? 0 : 1
-    `local_authority_import_check #{code} "#{message}"`
-  end
-
-  def self.update
-    fh = fetch_data
-    begin
-      new(fh).run
-    ensure
-      fh.close
-    end
-  end
-
-  def self.fetch_http_to_file(url)
-    fh = Tempfile.new(['local_authority_data', 'csv'])
-    fh.set_encoding('ascii-8bit')
-
-    uri = URI.parse(url)
-    response = Net::HTTP.get_response(uri)
-
-    if response.code != "200"
-      raise "HTTP Fetch failed [#{response.code}]: #{url}"
-    end
-
-    fh.write response.body
-
-    fh.rewind
-    fh.set_encoding('windows-1252', 'UTF-8')
-    fh
-  end
-
-  def initialize(fh)
-    @filehandle = fh
-  end
-
-  def run
-    CSV.new(@filehandle, headers: true).each do |row|
-      begin
-        process_row(row)
-      rescue => e
-        Rails.logger.error "Error #{e.class} processing row in #{self.class}\n#{e.backtrace.join("\n")}"
-        Airbrake.notify_or_ignore(e, :parameters => {:row => row})
-      end
-    end
-  end
-end
