--- test/unit/fixtures/local_contacts_sample.csv
@@ -1,4 +1,0 @@
-Name,Home page URL,Contact page URL,SNAC Code,Address Line 1,Address Line 2,Town,City,County,Postcode,Telephone Number 1 Description,Telephone Number 1,Telephone Number 2 Description,Telephone Number 2,Telephone Number 3 Description,Telephone Number 3,Fax,Main Contact Email,Opening Hours
-Adur District Council,http://www.adur.gov.uk,http://www.adur.gov.uk/contact-us/index.htm,45UB,Civic Centre,Ham Road,Shoreham-by-Sea,,West Sussex,BN43 6PR,Main switchboard,01273 263 000,,,,,01273 263 224,info@adur.gov.uk,
-Allerdale Borough Council,http://www.allerdale.gov.uk,http://www.allerdale.gov.uk/contact-or-find-us.aspx,16UB,Allerdale House,,Workington,,Cumbria,CA14 3YJ,,01900 702 702,,,,,01900 702 507,enquiries@allerdale.gov.uk,
-Basildon District Council,http://www.basildon.gov.uk/,http://www.basildon.gov.uk/contactus,22UB,The Basildon Centre,St. Martin's Square,Basildon,,Essex,SS14 1DL,Main switchboard,01268 533 333,Out of hours emergency,01268 286 622,,,01268 294 350,mailroom@basildon.gov.uk,
