--- test/unit/fixtures/local_interactions_county_sample.csv
@@ -1,2 +1,0 @@
-Authority Name,SNAC,LAid,Service Name,LGSL,LGIL,Service URL
-Buckinghamshire County Council,11,45,Apply for nursery school places,12,0,http://www.buckscc.gov.uk/bcc/early_years/choosing_a_place.page?
