--- test/unit/fixtures/local_interactions_encoding_samples.csv
@@ -1,4 +1,0 @@
-Authority Name,SNAC,LAid,Service Name,LGSL,LGIL,Service URL
-Adur District Council,45UB,1,Find out about  cooling tower registration  ,707,8,http://www.adur.gov.uk/licences/
-Adur District Council,45UB,1,Find out about details of council expenditure over £500,1465,8,http://www.adur.gov.uk/finance/payments-to-suppliers.htm
-Buckinghamshire County Council,11,45,Apply to join a library,438,0,http://buckinghamshire.spydus.co.uk/cgi-bin/spydus.exe/MSGTRN/OPAC/JOIN
