--- test/unit/fixtures/local_interactions_invalid_snac.csv
@@ -1,3 +1,0 @@
-Authority Name,SNAC,LAid,Service Name,LGSL,LGIL,Service URL
-Isles of Scilly,OOHF,166,Apply for a secondary school place,14,0,http://www.scilly.gov.uk/education/schools/default.htm
-Adur District Council,45UB,1,Find out about school holiday schemes,18,8,http://www.adur.gov.uk/education/index.htm
