--- test/unit/fixtures/local_interactions_ni_sample.csv
@@ -1,2 +1,0 @@
-Authority Name,SNAC,LAid,Service Name,LGSL,LGIL,Service URL
-Strabane District Council,95J,498,Apply for food premises registration,390,0,http://www.ukwelcomes.businesslink.gov.uk/bdotg/action/piplink?agency_id=498&service_id=5900010001&site=2000
