--- test/unit/fixtures/local_interactions_sample.csv
@@ -1,2 +1,0 @@
-Authority Name,SNAC,LAid,Service Name,LGSL,LGIL,Service URL
-Adur District Council,45UB,1,Find out about school holiday schemes,18,8,http://www.adur.gov.uk/education/index.htm
