--- test/unit/fixtures/mapit_response.json
@@ -1,1 +1,0 @@
-{"generation_high": 15, "codes": {"ons": "45UB", "gss": "E07000223", "unit_id": "16223"}, "name": "Adur District Council", "generation_low": 1, "country_name": "England", "country": "E", "type_name": "District council", "type": "DIS", "id": 2464, "parent_area": null}
